Thread-safety fix.
When we prepared the episode, our `Atomic` class had a `mutate` method on it, and no setter for `value`. In a last-minute decision before recording, we decided to go with a setter instead. As [Adam Sharp](https://twitter.com/sharplet) and Tobias Haeberle pointed out, the code contains a subtle thread-safety bug. ReactiveSwift contains the same bug, and Adam [opened an issue](https://github.com/ReactiveCocoa/ReactiveSwift/issues/269).